pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic"
version = "0.1.0"
description = "A comprehensive agentic AI framework implementing 21 design patterns"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Agentic Framework" }]
keywords = ["ai", "agents", "llm", "agentic", "patterns"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # LLM Providers
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    # HTTP / async
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
    # RAG / Vector store
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.7.0",
    # Observability
    "structlog>=24.2.0",
    "rich>=13.7.0",
    # Utilities
    "tenacity>=8.3.0",
    "jinja2>=3.1.4",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
]

[project.entry-points."agentic.plugins"]
# Built-in plugins registered here; third-party plugins add their own entry_points
# example = "agentic.plugins.example:ExamplePlugin"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
